power: qpnp-fg: queue work on system_power_efficient_wq
There doesn't seem to be any real dependency of scheduling these on
the cpu which scheduled them, so moving every *_delayed to the
power efficient wq save potential needlessly idle cpu wake ups
leaving the scheduler to decide the most appropriate cpus to wake up.

These are called very often so I wouldn't be susprised if the power
savings with this change were somewhat significant.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

# Conflicts:
#	drivers/power/qpnp-fg.c (reverted from commit 69af2ebb5ec72ed4e610a1de047eb94e3a4d135e)